budget_falcon/requirements.txt
boto3==1.38.21 ; python_version >= "3.13"
botocore==1.38.21 ; python_version >= "3.13"
cachetools==5.5.2 ; python_version >= "3.13"
certifi==2025.4.26 ; python_version >= "3.13"
charset-normalizer==3.4.2 ; python_version >= "3.13"
contourpy==1.3.2 ; python_version >= "3.13"
cycler==0.12.1 ; python_version >= "3.13"
fonttools==4.58.0 ; python_version >= "3.13"
google-api-core==2.24.2 ; python_version >= "3.13"
google-api-python-client==2.169.0 ; python_version >= "3.13"
google-auth-httplib2==0.2.0 ; python_version >= "3.13"
google-auth==2.40.2 ; python_version >= "3.13"
googleapis-common-protos==1.70.0 ; python_version >= "3.13"
httplib2==0.22.0 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
jmespath==1.0.1 ; python_version >= "3.13"
kiwisolver==1.4.8 ; python_version >= "3.13"
matplotlib==3.10.3 ; python_version >= "3.13"
numpy==2.2.6 ; python_version >= "3.13"
packaging==25.0 ; python_version >= "3.13"
pillow==11.2.1 ; python_version >= "3.13"
proto-plus==1.26.1 ; python_version >= "3.13"
protobuf==6.31.0 ; python_version >= "3.13"
pyasn1-modules==0.4.2 ; python_version >= "3.13"
pyasn1==0.6.1 ; python_version >= "3.13"
pyparsing==3.2.3 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
pytz==2025.2 ; python_version >= "3.13"
pyyaml==6.0.2 ; python_version >= "3.13"
requests==2.32.3 ; python_version >= "3.13"
rsa==4.2 ; python_version >= "3.13"
s3transfer==0.12.0 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
slack-sdk==3.35.0 ; python_version >= "3.13"
uritemplate==4.1.1 ; python_version >= "3.13"
urllib3==2.4.0 ; python_version >= "3.13"
